fix(phase1): FailoverManager delegates leader switch to Raft consensus
- PartitionFailoverController no longer directly mutates current_leader
- Added ConsensusTransferCallback; production requires it to be wired
- Without callback, failover is rejected with 'manual intervention required'

BLOCKER fix: Distributed Correctness #3

+++ src/sst/zone_encoder.cc
@@ -41,7 +41,7 @@ void VarintCodec::Encode(uint64_t value, std::string& dst) {
 
 size_t VarintCodec::EncodeSigned(int64_t value, char* buf) {
   // ZigZag 编码：将符号位移到最低位
-  uint64_t encoded = (value << 1) ^ (value >> 63);
+  uint64_t encoded = (static_cast<uint64_t>(value) << 1) ^ static_cast<uint64_t>(value >> 63);
   return Encode(encoded, buf);
 }
